+++ voxygen/src/menu/char_selection/mod.rs
@@ -117,10 +117,11 @@ impl PlayState for CharSelectionState {
                         mainhand,
                         offhand,
                         body,
+                        start_site,
                     } => {
                         self.client
                             .borrow_mut()
-                            .create_character(alias, mainhand, offhand, body);
+                            .create_character(alias, mainhand, offhand, body, start_site);
                     },
                     ui::Event::EditCharacter {
                         alias,