+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   will only lock the camera zoom while movement and combat inputs are also being pressed.
 - Custom spots can be added without recompilation (only ron and vox files)
 - Setting in userdata/server/server_config/settings.ron that controls the length of each day/night cycle.
+- Starting site can now be chosen during character creation
 
 ### Changed
 - Bats move slower and use a simple proportional controller to maintain altitude